+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 1.2.0 Release (25.11.2020)
+* Added NumberUtils#isLong method to check if given string is primitive type long.
+* Added NumberUtils#isShort method to check if given string is primitive type short.
 * Changes on scoreboard system:
   * Fixed colors in Minecraft 1.16 and higher versions
   * Fixed not working in Minecraft 1.16.4